added JEST tests
