+++ package.json
@@ -30,6 +30,7 @@
     "file-loader": "^6.0.0",
     "geonames.js": "^2.3.0",
     "html-webpack-plugin": "^4.0.3",
+    "jest": "^25.4.0",
     "mini-css-extract-plugin": "^0.9.0",
     "node": "^13.12.0",
     "node-fetch": "^2.6.0",
