+++ src/client/js/request.test.js
@@ -1,0 +1,6 @@
+describe('http requests', () => {
+    it('should make get and post connections', async () => {
+        const data = await expect (Promise.resolve('response'));
+        expect(data).toEqual(expect.anything());
+    })
+})
