+++ src/client/js/setImg.js
@@ -1,0 +1,4 @@
+export const setImg = (pixImg) => {
+    const imgLocation = document.getElementById("image");
+    imgLocation.setAttribute("src", pixImg);
+  }
