+++ src/server/server.test.js
@@ -1,0 +1,6 @@
+
+const city = {found: true};
+
+test('should find a city value', () => {
+    expect(city.found).toBeTruthy();
+})
