+++ src/styles/styles.scss
@@ -1,14 +1,24 @@
-.wrapper {
-  display: grid;
-  grid-template-columns: repeat(auto-fill, 200px);
-  grid-template-rows: repeat(auto, 1fr); 
+.paragraph {
+  color: #c1f3d0;
+  font-family: Luminari, fantasy;
+  font-size: 24px;
+  margin: 20px;
+  text-align: center;
 }
 
-.paragraph {
-  color: white;
-  font-family: -apple-system, BlinkMacSystemFont, sans-serif;
-  font-size: 18px;
+.results {
+  display: flex;
+  flex-direction: column;
+  flex-wrap: wrap;
+  justify-content: center;
+}
+
+h2 {
+  color: #c1f3d0;
+  font-family: Luminari, fantasy;
+  font-size: 24px;
   margin: 20px;
+  text-align: center;
 }
 
 figure {
@@ -18,73 +28,52 @@ figure {
 }
 
 body {
-  background-image: url('../media/background.jpg');
-  background-attachment: fixed;
-  background-repeat: cover;
-  background-repeat: no-repeat;
-  background-size: 100%;
+  height: auto;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  background: url('../media/background.jpg') center no-repeat fixed;
+  background-size: cover;
 }
 
-.container {
-  display: flex;
-  // grid-column: 3 / 6;
-  flex-direction: column;
-  flex-wrap: wrap;
-  justify-content: center;
-  justify-self: center;
-}
-.results {
-  display: flex;
-  flex-direction: column;
-  justify-self: center;
+.res-txt {
+  color: #c1f3d0;
+  font-family: Luminari, fantasy;
+  font-size: 20px;
+  margin: 20px;
+  text-align: center;
 }
 
 #cityValue {
   display: flex;
-  flex-direction: row;
   justify-items: center;
-  background: #f23557;
+  justify-content: center;
 }
 
 #summaryRes {
-  display: flex;
-  flex-direction: row;
-  justify-items: center;
-  background: #f23557;
+  box-sizing: border-box;
+  padding: 10px;
 }
 
 #iconRes {
-  display: flex;
-  flex-direction: row;
-  justify-items: center;
-  background: #f23557;
+  align-items: center;
+  box-sizing: border-box;
+  padding: 10px;
 }
 
 #imgRes {
   display: flex;
-  flex-direction: row;
+  flex-direction: column;
   justify-items: center;
-  background: #f23557;
 }
 
-// #app {
-//   display: grid;
-//   grid-auto-rows: minmax(150px, auto);
-//   /* grid-template-columns:  */
-//   height: 100vh;
-//   background: #f23557;
-//   color: #f0d43a;
-//   justify-content: center;
-//   grid-gap: 1em;
-//   font-size: 20px;
-//   font-family: "Oswald", sans-serif;
-// }
-
-.form {
-  display: flex;
-  flex-direction: row;
-  flex-wrap: wrap;
-  margin-bottom: 20px;
+canvas {
+  padding-left: 0;
+  padding-right: 0;
+  margin-left: auto;
+  margin-right: auto;
+  display: block;
 }
 
 .entry {
@@ -100,32 +89,41 @@ body {
 }
 /*  Basic Styling To Override Default For Basic HTML Elements */
 
-#city {
-  margin-bottom: 20px;
-}
-
 .input {
   display: block;
   height: 60px;
   width: 320px;
   background: #005a5a;
-  color: #baf4ce;
+  color: #c1f3d0;
   font-size: 20px;
-  font-family: "Oswald", sans-serif;
+  font-family: Luminari, fantasy;
   border: none;
+  text-align: center;
+  align-self: center;
 }
 
 button {
-  width: 400px;
-  height: 75px;
-  background: #3b4a6b;
-  color: #f0d43a;
-  font-size: 26px;
-  font-family: "Oswald", sans-serif;
-  border: none;
+  width: 150px;
+  height: 50px;
+  margin-top: 30px;
+  position: relative;
+  margin-left: 90px;
+  color: #c1f3d0;
+  background: #005a5a;
+  font-size: 18px;
+  font-weight: bold;
+  font-family: Bradley Hand, cursive;
+  border-radius: 5px;
   box-shadow: 2px 4px 5px #444;
   cursor: pointer;
 }
+
+button:hover{
+  color: #005a5a;
+  background: #baf4ce;
+  transition: all 0.3s ease;
+}
+
 h1 {
   font-size: 36px;
 }
@@ -141,16 +139,6 @@ h1 {
   padding-bottom: 20px;
 }
 
-#submit {
-  margin-bottom: 30px;
-}
-
-textarea {
-  background: #22b2da;
-  color: #f0d43a;
-  font-size: 20px;
-  font-family: "Oswald", sans-serif;
-}
 
 /* Reset style for HTML element on active to have no border*/
 input:focus,
